requirements.txt
Django==3.2
djangorestframework==3.12.4
requests==2.25.1
beautifulsoup4==4.11.1
